backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
environs[django]==11.0.0
mssql-django==1.5
pyodbc==5.1.0
